requirements.txt
# ══════════════════════════════════════════
# PRISM-LLM Dependencies
# ══════════════════════════════════════════

# --- Core ML ---
torch>=2.1.0
transformers>=4.36.0
sentence-transformers>=2.2.2
accelerate>=0.25.0
peft>=0.7.0

# --- LLM Backends & Orchestration ---
anthropic>=0.18.0
openai>=1.6.0
langchain>=0.1.0
langgraph>=0.0.20

# --- Retrieval & NLP ---
rank-bm25>=0.2.2
faiss-cpu>=1.7.4
nltk>=3.8.1
scikit-learn>=1.3.0

# --- PDF & Document Processing ---
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0

# --- Data Validation & Schemas ---
pydantic>=2.5.0
jsonschema>=4.20.0
pandas>=2.1.0
numpy>=1.25.0

# --- Database Connectors ---
biopython>=1.82
pybliometrics>=3.5.0
requests>=2.31.0

# --- Evaluation & Statistics ---
scipy>=1.11.0
statsmodels>=0.14.0

# --- Visualization ---
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# --- Vector Database ---
qdrant-client>=1.7.0

# --- Utilities ---
pyyaml>=6.0.1
tqdm>=4.66.0
loguru>=0.7.0
python-dotenv>=1.0.0
python-Levenshtein>=0.23.0

# --- Testing ---
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
